Add files via upload
The LED have been powered by Arduino UNO (Board). It contains a code which uploaded to the board. And once it simulated LED Start's blinking like a traffic light. In this 15 Second will for Red Light 6 Second for Yellow Light (In my Project Blue) and 20 Second for Green Light.
Traffic signals are designed to ensure an orderly flow of traffic, provide an opportunity for pedestrians or vehicles to cross an intersection and help reduce the number of conflicts between vehicles entering intersections from different directions.  Under this module, the traffic signal control system takes into account the vehicle demand volume using the road-embedded detector and adjusts the green light accordingly.in case of jammed roads, it adjusts the timings accordingly to step UP the traffic flows. The primary function of any traffic signal is to assign right-of-way to conflicting movements of traffic at an intersection. This is done by permitting conflicting streams of traffic to share the same intersection by means of time separation. Traffic light control systems are widely used to monitor and control the flow of automobiles through the junction of many roads. They aim to realize smooth motion of cars in the transportation routes.